% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modaw.R
\name{auto_var}
\alias{auto_var}
\title{Write scripts of latent variables and/or residual correlations of items}
\usage{
auto_var(variable_name, item_names = NULL, resi_cor = TRUE,
  item = NULL, data = NULL)
}
\arguments{
\item{variable_name}{Charater. Name of the variable.}

\item{item_names}{Charater or charater vector. Names of the items.}

\item{resi_cor}{Logical. Whether include residual correlations.}

\item{item}{Integer. The column number of items in a data frame}

\item{data}{The input data frame}
}
\value{
Characters.
}
\description{
Automatically generate the scripts of latent variables and/or residual correlations of items used for lavaan model
}
\examples{
## item names
item_names <- c("em1.an1", "em1.an2", "em1.an3", "em2.an1", "em2.an2", "em2.an3",
                "em3.an1", "em3.an2", "em3.an3")

## get the scripts used for lavaan model by entering the item names
auto_var(variable_name = "E.A", item_names = item_names)

## creat a data frame with 9 items
df <- data.frame(
  em1.an1 = c(3, 4, 5, 1),
  em1.an2 = c(1, 2, 5, 3),
  em1.an3 = c(1, 4, 5, 3),
  em2.an1 = c(1, 2, 5, 5),
  em2.an2 = c(5, 2, 5, 5),
  em2.an3 = c(5, 1, 4, 5),
  em3.an1 = c(4, 1, 4, 1),
  em3.an2 = c(2, 1, 1, 1),
  em3.an3 = c(2, 3, 2, 1),
  stringsAsFactors = FALSE
)

## get the scripts used for lavaan model by entering the column number of the data frame
auto_var(variable_name = "E.A", item = 1:9, data = df)

}
